chore: address review nit in recovery tracker mapping
Agent-Logs-Url: https://github.com/rfivesix/hypertrack/sessions/7fe63cf3-14ec-49d3-a2a7-08f9dd6bbdcc

+++ lib/screens/analytics/recovery_tracker_screen.dart
@@ -252,8 +252,7 @@ class _RecoveryTrackerScreenState extends State<RecoveryTrackerScreen> {
                       final muscleName = muscle.muscleGroup;
                       final state = muscle.state;
                       final stateColor = _stateColor(context, state);
-                      final hours =
-                          muscle.hoursSinceLastSignificantLoad.toDouble().round();
+                      final hours = muscle.hoursSinceLastSignificantLoad.round();
                       final highFatigue = muscle.highSessionFatigue;
                       final eqSets = muscle.lastEquivalentSets;
                       final recoveringUpper = muscle.recoveringUpperHours;
